架构：完整支持多架构 Linux (x86_64/ARM32/ARM64/RISC-V64)
## 核心改进

### 1. 类型系统增强 (osal_types.h)
- 添加 RISC-V 架构检测 (__riscv && __riscv_xlen)
- 添加 ARM32 架构检测 (__arm__)
- 添加完整字节序支持：
  * 自动检测大小端 (OSAL_LITTLE_ENDIAN/OSAL_BIG_ENDIAN)
  * 字节序转换宏 (OSAL_HTONS/HTONL/HTONLL)
  * 使用 GCC/Clang 内置函数优化 (__builtin_bswap16/32/64)

### 2. 原子操作标准化
- OSAL 层：osal_task.c 使用 _Atomic int32_t 替代 atomic_int
- HAL 层：hal_can.c 使用 _Atomic uint32_t 替代 atomic_uint
- 统一使用固定大小原子类型，保证跨架构一致性

### 3. 指针转换安全性
- hal_spi.c: 使用 uintptr_t 替代 unsigned long (2处)
- hal_i2c.c: 使用 uintptr_t 替代 unsigned long (2处)
- 修复 32位/64位平台指针转换问题

### 4. 构建系统多架构支持 (CMakeLists.txt)
- 新增平台：arm32-linux, arm64-linux, riscv64-linux
- 架构特定编译选项：
  * ARM32: -march=armv7-a -mfpu=neon-vfpv4 -mfloat-abi=hard
  * ARM64: -march=armv8-a
  * RISC-V64: -march=rv64imafdc -mabi=lp64d
- 保留 generic-linux 兼容性（已标记为 deprecated）

### 5. 编译脚本增强 (build.sh)
- 新增 -a/--arch 参数支持多架构选择
- 支持架构：native, x86_64, arm32, arm64, riscv64
- 自动检测交叉编译工具链
- 工具链缺失时提供安装提示
- 显示目标架构和构建类型信息

### 6. 文档更新
- CLAUDE.md: 添加多架构编译说明和交叉编译示例
- README.md: 添加交叉编译快速入门和工具链安装指南
- 说明架构特性：固定大小类型、字节序、原子操作

## 技术细节

**支持的架构**：
- x86_64: Intel/AMD 64位 (小端)
- ARM32: ARMv7-A (小端, 硬浮点)
- ARM64: ARMv8-A (小端)
- RISC-V 64: RV64IMAFDC (小端, lp64d ABI)

**交叉编译工具链**：
- ARM32: arm-linux-gnueabihf-gcc
- ARM64: aarch64-linux-gnu-gcc
- RISC-V64: riscv64-linux-gnu-gcc

**编译示例**：
```bash
./build.sh -a arm64           # ARM64 Release
./build.sh -a riscv64 -d      # RISC-V64 Debug
```

## 测试验证

- 代码已通过静态分析
- 类型系统保证跨架构一致性
- 字节序转换使用编译器内置函数优化
- 原子操作使用 C11 标准，所有架构支持

## 影响范围

- 修改文件：9个
- 新增代码：317行
- 删除代码：19行
- 向后兼容：完全兼容现有代码

+++ hal/src/linux/hal_i2c.c
@@ -115,7 +115,7 @@ int32_t HAL_I2C_Write(hal_i2c_handle_t handle, uint16_t slave_addr,
         return OSAL_ERR_GENERIC;
 
     /* 设置从设备地址 (参考: Linux内核文档 i2c-dev.txt) */
-    ret = OSAL_ioctl(impl->fd, I2C_SLAVE, (void *)(unsigned long)slave_addr);
+    ret = OSAL_ioctl(impl->fd, I2C_SLAVE, (void *)(uintptr_t)slave_addr);
     if (ret < 0)
     {
         LOG_ERROR("HAL_I2C", "Failed to set slave address 0x%02X: %s",
@@ -156,7 +156,7 @@ int32_t HAL_I2C_Read(hal_i2c_handle_t handle, uint16_t slave_addr,
         return OSAL_ERR_GENERIC;
 
     /* 设置从设备地址 */
-    ret = OSAL_ioctl(impl->fd, I2C_SLAVE, (void *)(unsigned long)slave_addr);
+    ret = OSAL_ioctl(impl->fd, I2C_SLAVE, (void *)(uintptr_t)slave_addr);
     if (ret < 0)
     {
         LOG_ERROR("HAL_I2C", "Failed to set slave address 0x%02X: %s",
